♻️ Seperate UpperBar components
- UpperBar 내의 요소들을 GoBackButton, SearchBar 컴포넌트로 분리해줌
